# Implement 3D shadows in the GL Compatibility renderer