… drivers/gles3/rasterizer_scene_gles3.cpp
… drivers/gles3/shaders/scene.glsl
… drivers/gles3/storage/light_storage.cpp
… servers/rendering/renderer_rd/storage_rd/light_storage.h