First pass at implementing shadows in 3D in the GL Compatibility renderer. This was way more work than I thought. 

Implementation based on https://github.com/godotengine/godot-proposals/issues/3959

### Main approach

Particularly this uses a multipass approach to lighting so that we can benefit from prefetch when using hard shadows. In practice this means:
1. All lights without shadows are rendered during base pass
2. Each light with a shadow requires drawing the object again
3. Lights with shadows are blended in sRGB space instead of linear, so the colors blow out very quickly (i.e. lights with shadows look like they do in the Godot 3.x GLES2 renderer) (Lights without shadows look normal)

### Concerns

In theory this approach should be as fast as possible for mobile devices (the speed benefit is less observable on desktop). However, I have a few concerns:
1. In order to avoid lights flickering between Linear and sRGB blending, shadowed lights _always_ have to be in a separate pass. In https://github.com/godotengine/godot-proposals/issues/3959 reduz assumed we could render one positional and one directional light in the base pass, but in practice this looks really bad. In theory, many games could stick to the limitation of having at most one directional shadow and one positional shadow per object to maximize performance, but with this implementation a single positional shadow and directional shadow would result in 3 draw passes!
2. Prefetch only applies to texture samples taken directly from interpolaters, so all the PCF5 and PCF13 samples don't benefit.

### Comments

Omni light shadows are still a bit borked. I do not want to spend the time debugging them until we have a discussion about the overall approach. In particular, I need to do some performance testing before committing to this general approach. My gut tells me the performance gains from doing multipass in terms of prefetchare outweighed by the performance losses from having to render each object multiple times.