@Calinou 2023-05-26
> My gut tells me the performance gains from doing multipass in terms of prefetchare outweighed by the performance losses from having to render each object multiple times.

In practice, we can't expect every mobile project to be using fully baked lightmaps, so I think improving performance of rendering multiple real-time lights with shadows is key.