@piramidamsc 2023-06-07
Directional Light doesn't work properly for me, I'm using your custom godot 4.1 dev build