@Calinou 2023-07-09
> Does this fix the longstanding issue where shadows ignore shader parameter POINT_SIZE for point meshes Mesh.PRIMITIVE_POINTS?

Please open a separate issue with a minimal reproduction project attached.