@reduz 2023-08-14
@clayjohn 
> In order to avoid lights flickering between Linear and sRGB blending, shadowed lights always have to be in a separate pass. In https://github.com/godotengine/godot-proposals/issues/3959 reduz assumed we could render one positional and one directional light in the base pass, but in practice this looks really bad.

I can see why. That said, I would still try to, at least, do the directional light in a single pass merged with the non shadowed lights. Directional + non shadowed lights is the most common way to do lighting on mobile.

Also shadow filter is ok and expected to be non-pcf on mobile by default (only enable linear interpolation so it does hardware accelerated PCF4, which is per spec on GLES3).