@reduz 2023-08-14
@Calinou I am surprised performance is not similar to Godot GLES2 to be honest. We really should aim for as closest as that if possible.