@clayjohn 2023-08-14
> @clayjohn
> 
> > In order to avoid lights flickering between Linear and sRGB blending, shadowed lights always have to be in a separate pass. In [godotengine/godot-proposals#3959](https://github.com/godotengine/godot-proposals/issues/3959) reduz assumed we could render one positional and one directional light in the base pass, but in practice this looks really bad.
> 
> I can see why. That said, I would still try to, at least, do the directional light in a single pass merged with the non shadowed lights. Directional + non shadowed lights is the most common way to do lighting on mobile, so this as single pass is ideal.
> 
> I know this would break if you have more than one directional light, but we could have a setting `allow_multiple_directional_lights` on compatibility that works as you describe _only_ if user enables it.
> 
> Also shadow filter is ok and expected to be non-pcf on mobile by default (only enable linear interpolation so it does hardware accelerated PCF4, which is per spec on GLES3).

I already have a plan. I will move the calculation of the additional lights to after the tonemapping step in the shader. That way, we can have the first light with shadows calculated in the original pass, but still in sRGB space. Should be enough to improve performance a little bit and ensure we get the best performance possible when using only a few lights.