@reduz 2023-09-28
I still wonder why the GLES2 version of Godot is faster, what do we do different there?