@clayjohn 2023-09-28
> I still wonder why the GLES2 version of Godot 3 is faster, what do we do different there?

Those measurements were taken with the old version of this PR. Back then we did an additional pass per light with shadows. Now the base pass includes one shadowed light so it is much faster