@Calinou 2023-09-28
I've run benchmarks again on the latest revision of this PR (same hardware as my prior test, also in 3840×2160):


3.5.2 GLES3 | 3.5.2 GLES2 | `master` Forward+ | `master` Mobile | `master` Compatibility<br> *(no shadows)* | 4.x Compatibility<br> *(this PR)*
-|-|-|-|-|-
1705 FPS (0.59 mspf) | 2550 FPS (0.39 mspf) | 1249 FPS (0.80 mspf) | 1247 FPS (0.80 mspf) | 2890 FPS (0.35 mspf) | 1828 FPS (0.55 mspf)

Directional shadow rendering still appears broken in the running project, even though it works in the editor:

![Screenshot_20230928_143329 webp](https://github.com/godotengine/godot/assets/180032/565eb40e-609c-4184-a6ee-86eb7b2fa7f1)

In comparison, this is how the scene looks like in Forward+:

![Screenshot_20230928_143932 webp](https://github.com/godotengine/godot/assets/180032/6b998b65-af3e-444c-a564-56b4c133f533)

Visual comparison in the editor:

Forward+ | Mobile | Compatibility
-|-|-
![Screenshot_20230928_143726 webp](https://github.com/godotengine/godot/assets/180032/1c66ea5a-6304-4488-a858-ed4f8a77e779) | ![Screenshot_20230928_143744 webp](https://github.com/godotengine/godot/assets/180032/ff6e3739-bde9-4371-8357-fab2b5da21b5) | ![Screenshot_20230928_143751 webp](https://github.com/godotengine/godot/assets/180032/d346cbec-ccee-4f09-9af9-3e0e16689aab)