@clayjohn 2023-09-29
Just noting a comment from the top post:

> There is still one outstanding bug where shadow atlas textures are getting out of date with multiple shadowed lights in the scene.

I'll investigate that outstanding bug and any new reports ASAP